Update messaging, add alive status to get_stats fn

+++ R/interact.R
@@ -61,7 +61,6 @@ get_stats <- function() {
   if (bp$experience$level == 2L) level_text <- internal$constants$level_names$lvl_2
   if (bp$experience$level == 3L) level_text <- internal$constants$level_names$lvl_3
   if (bp$experience$level == 4L) level_text <- internal$constants$level_names$lvl_4
-  if (!bp$meta$alive) level_text <- "unalive"
 
   empty_happy  <- rep("\U025A0", bp$status$happy)
   empty_hungry <- rep("\U025A0", bp$status$hungry)
@@ -71,38 +70,21 @@ get_stats <- function() {
   filled_hungry <- rep("\U025A1", 5 - bp$status$hungry)
   filled_dirty  <- rep("\U025A1", 5 - bp$status$dirty)
 
-  if (bp$status$happy == 0L) {
-    warn_happy  <- " !"
-  } else {
-    warn_happy <-  "  "
-  }
-
-  if (bp$status$hungry == 5L) {
-    warn_hungry <- " !"
-  } else {
-    warn_hungry <- "  "
-  }
-
-  if (bp$status$dirty == 5L) {
-    warn_dirty  <- " !"
-  } else {
-    warn_dirty  <- "  "
-  }
-
   message(
     "Characteristics",
     "\n  Name:    ", bp$characteristics$name,
     "\n  Species: ", bp$characteristics$species,
     "\n  Age:     ", bp$characteristics$age,
-    "\n  Level:   ", bp$experience$level, paste0(" (", level_text, ")")
+    "\n  Level:   ", bp$experience$level, paste0(" (", level_text, ")"),
+    "\n  Alive:   ", bp$meta$alive, ifelse(!bp$meta$alive, " :(", "")
   )
 
   if (bp$meta$alive) {
     message(
       "Status",
-      "\n  Happy:   ", empty_happy,  filled_happy,  warn_happy,
-      "\n  Hungry:  ", empty_hungry, filled_hungry, warn_hungry,
-      "\n  Dirty:   ", empty_dirty,  filled_dirty,  warn_dirty
+      "\n  Happy:   ", empty_happy,  filled_happy,  ifelse(bp$status$happy == 0L,  " !", "  "),
+      "\n  Hungry:  ", empty_hungry, filled_hungry, ifelse(bp$status$hungry == 5L, " !", "  "),
+      "\n  Dirty:   ", empty_dirty,  filled_dirty,  ifelse(bp$status$dirty == 5L,  " !", "  ")
     )
   }
 
@@ -153,6 +135,13 @@ see_pet <- function() {
 play <- function() {
 
   bp <- .check_and_update()
+
+  if (!bp$meta$alive) {
+    stop(
+      bp$characteristics$name, " is unalive! Can't play.",
+      call. = FALSE
+    )
+  }
 
   if (bp$status$happy == 5) {
     stop(
@@ -257,6 +246,13 @@ feed <- function() {
 
   bp <- .check_and_update()
 
+  if (!bp$meta$alive) {
+    stop(
+      bp$characteristics$name, " is unalive! Can't feed.",
+      call. = FALSE
+    )
+  }
+
   if (bp$status$hungry == 0) {
     stop(
       "Hunger is already at the minimum value! Can't feed.",
@@ -283,6 +279,13 @@ feed <- function() {
 clean <- function() {
 
   bp <- .check_and_update()
+
+  if (!bp$meta$alive) {
+    stop(
+      bp$characteristics$name, " is unalive! Can't clean.",
+      call. = FALSE
+    )
+  }
 
   if (bp$status$dirty == 0) {
     stop(
@@ -315,25 +318,59 @@ release <- function() {
 
   bp <- .check_and_update()
 
-  answer_a <-
-    readline(paste0("Really release ", bp$characteristics$name, "? y/n: "))
+  if (bp$meta$alive) {
 
-  if (substr(tolower(answer_a), 1, 1) == "y") {
+    answer_a <-
+      readline(paste0("Really release ", bp$characteristics$name, "? y/n: "))
 
-    answer_b <- readline("Are you sure? y/n: ")
+    if (substr(tolower(answer_a), 1, 1) == "y") {
 
-    if (substr(tolower(answer_b), 1, 1) == "y") {
+      answer_b <- readline("Are you sure? y/n: ")
 
-      blueprint_path <- tools::R_user_dir("tamRgo", which = "data")
-      file.remove(file.path(blueprint_path, "blueprint.rds"))
-      message(bp$characteristics$name, " was set free!")
+      if (substr(tolower(answer_b), 1, 1) == "y") {
+
+        blueprint_path <- tools::R_user_dir("tamRgo", which = "data")
+        file.remove(file.path(blueprint_path, "blueprint.rds"))
+        message(bp$characteristics$name, " was set free!")
+
+      } else {
+        message(bp$characteristics$name, " was not released.")
+      }
 
     } else {
       message(bp$characteristics$name, " was not released.")
     }
 
-  } else {
-    message(bp$characteristics$name, " was not released.")
+  }
+
+  if (!bp$meta$alive) {
+
+    answer_a <-
+      readline(
+        paste0(
+          bp$characteristics$name, " is unalive! Can't release. ",
+          "Delete blueprint? y/n: "
+          )
+        )
+
+    if (substr(tolower(answer_a), 1, 1) == "y") {
+
+      answer_b <- readline("Are you sure? y/n: ")
+
+      if (substr(tolower(answer_b), 1, 1) == "y") {
+
+        blueprint_path <- tools::R_user_dir("tamRgo", which = "data")
+        file.remove(file.path(blueprint_path, "blueprint.rds"))
+        message(bp$characteristics$name, "'s blueprint was deleted!")
+
+      } else {
+        message(bp$characteristics$name, "'s blueprint was not deleted")
+      }
+
+    } else {
+      message(bp$characteristics$name, "'s blueprint was not deleted")
+    }
+
   }
 
 }
